// src/interfaces/students.ts
export interface Student {
  tipoIdentificacion?: number
  numeroIdentificacion?: number
  estado?: string
  nombres: string
  apellidos: string
  celular: number
  correo: string
  linkedin: string
  github: string
}

export interface ApiStudent {
  estudiante_apellidos: string
  estudiante_celular: string
  estudiante_correo: string
  estudiante_estado: string
  estudiante_fechaCreacion: string
  estudiante_github: string
  estudiante_id: number
  estudiante_linkedin: string
  estudiante_nombres: string
}